+++ R/cormatrix_to_excel.R
@@ -47,162 +47,162 @@ cormatrix_to_excel <- function(data,
                                print.mat = TRUE,
                                ...) {
   if (missing(filename)) {
-    stop("Argument 'filename' required (as per CRAN policies)")
+    insight::format_error("Argument 'filename' required (as per CRAN policies).")
   }
 
   insight::check_if_installed("openxlsx2")
 
   # create correlation matrix with p values
-  cm <- correlation::correlation(data, ...)
-  cm <- summary(cm, redundant = TRUE)
-  all.columns <- 2:(ncol(cm))
+  cormatrix <- correlation::correlation(data, ...)
+  cormatrix <- summary(cormatrix, redundant = TRUE)
+  all.columns <- 2:(ncol(cormatrix))
   if (isTRUE(print.mat)) {
-    print(cm)
+    print(cormatrix)
   }
-  pf <- attr(cm, "p")
+  p_val <- attr(cormatrix, "p")
 
   # Define colours
-  style_gray <- c(rgb = "C1CDCD")
-  style_black <- c(rgb = "000000")
-  style_pink <- c(rgb = "FBCAC0")
-  style_peach <- c(rgb = "F79681")
-  style_red <- c(rgb = "F65534")
-  style_lightblue <- c(rgb = "97FFFF")
-  style_midblue <- c(rgb = "0AF3FF")
-  style_darkblue <- c(rgb = "00BFFF")
-  style_green1 <- c(rgb = "698B22")
-  style_green2 <- c(rgb = "9ACD32")
-  style_green3 <- c(rgb = "B3EE3A")
+  style_gray <- openxlsx2::wb_colour(hex = "C1CDCD")
+  style_black <- openxlsx2::wb_colour(hex = "000000")
+  style_pink <- openxlsx2::wb_colour(hex = "FBCAC0")
+  style_peach <- openxlsx2::wb_colour(hex = "F79681")
+  style_red <- openxlsx2::wb_colour(hex = "F65534")
+  style_lightblue <- openxlsx2::wb_colour(hex = "97FFFF")
+  style_midblue <- openxlsx2::wb_colour(hex = "0AF3FF")
+  style_darkblue <- openxlsx2::wb_colour(hex = "00BFFF")
+  style_green1 <- openxlsx2::wb_colour(hex = "698B22")
+  style_green2 <- openxlsx2::wb_colour(hex = "9ACD32")
+  style_green3 <- openxlsx2::wb_colour(hex = "B3EE3A")
 
   # Colours
   gray_style <- openxlsx2::create_dxfs_style(
-    bgFill = style_gray,
+    bg_fill = style_gray,
     font_color = style_black,
-    numFmt = "#.#0 _*_*_*"
+    num_fmt = "#.#0 _*_*_*"
   )
 
   p_style <- openxlsx2::create_dxfs_style(
-    bgFill = "",
+    bg_fill = "",
     font_color = style_black,
-    numFmt = "#.##0 _*_*_*"
+    num_fmt = "#.##0 _*_*_*"
   )
   p_style1 <- openxlsx2::create_dxfs_style(
-    bgFill = style_green1,
+    bg_fill = style_green1,
     font_color = style_black,
-    numFmt = "#.##0 _*_*_*"
+    num_fmt = "#.##0 _*_*_*"
   )
   p_style2 <- openxlsx2::create_dxfs_style(
-    bgFill = style_green2,
+    bg_fill = style_green2,
     font_color = style_black,
-    numFmt = "#.##0 _*_*_*"
+    num_fmt = "#.##0 _*_*_*"
   )
   p_style3 <- openxlsx2::create_dxfs_style(
-    bgFill = style_green3,
+    bg_fill = style_green3,
     font_color = style_black,
-    numFmt = "#.##0 _*_*_*"
+    num_fmt = "#.##0 _*_*_*"
   )
 
   # no star
   no_star <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 _*_*_*",
+    num_fmt = "#.#0 _*_*_*",
     font_color = style_black,
-    bgFill = ""
+    bg_fill = ""
   )
 
   # one star
   one_star_pink <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*_*_*",
+    num_fmt = "#.#0 \\*_*_*",
     font_color = style_black,
-    bgFill = style_pink
+    bg_fill = style_pink
   )
   one_star_peach <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*_*_*",
+    num_fmt = "#.#0 \\*_*_*",
     font_color = style_black,
-    bgFill = style_peach
+    bg_fill = style_peach
   )
   one_star_red <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*_*_*",
+    num_fmt = "#.#0 \\*_*_*",
     font_color = style_black,
-    bgFill = style_red
+    bg_fill = style_red
   )
   one_star_lightblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*_*_*",
+    num_fmt = "#.#0 \\*_*_*",
     font_color = style_black,
-    bgFill = style_lightblue
+    bg_fill = style_lightblue
   )
   one_star_midblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*_*_*",
+    num_fmt = "#.#0 \\*_*_*",
     font_color = style_black,
-    bgFill = style_midblue
+    bg_fill = style_midblue
   )
   one_star_darkblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*_*_*",
+    num_fmt = "#.#0 \\*_*_*",
     font_color = style_black,
-    bgFill = style_darkblue
+    bg_fill = style_darkblue
   )
 
   # two stars
   two_stars_pink <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*_*",
+    num_fmt = "#.#0 \\*\\*_*",
     font_color = style_black,
-    bgFill = style_pink
+    bg_fill = style_pink
   )
   two_stars_peach <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*_*",
+    num_fmt = "#.#0 \\*\\*_*",
     font_color = style_black,
-    bgFill = style_peach
+    bg_fill = style_peach
   )
   two_stars_red <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*_*",
+    num_fmt = "#.#0 \\*\\*_*",
     font_color = style_black,
-    bgFill = style_red
+    bg_fill = style_red
   )
   two_stars_lightblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*_*",
+    num_fmt = "#.#0 \\*\\*_*",
     font_color = style_black,
-    bgFill = style_lightblue
+    bg_fill = style_lightblue
   )
   two_stars_midblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*_*",
+    num_fmt = "#.#0 \\*\\*_*",
     font_color = style_black,
-    bgFill = style_midblue
+    bg_fill = style_midblue
   )
   two_stars_darkblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*_*",
+    num_fmt = "#.#0 \\*\\*_*",
     font_color = style_black,
-    bgFill = style_darkblue
+    bg_fill = style_darkblue
   )
 
   # three stars
   three_stars_pink <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*\\*",
+    num_fmt = "#.#0 \\*\\*\\*",
     font_color = style_black,
-    bgFill = style_pink
+    bg_fill = style_pink
   )
   three_stars_peach <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*\\*",
+    num_fmt = "#.#0 \\*\\*\\*",
     font_color = style_black,
-    bgFill = style_peach
+    bg_fill = style_peach
   )
   three_stars_red <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*\\*",
+    num_fmt = "#.#0 \\*\\*\\*",
     font_color = style_black,
-    bgFill = style_red
+    bg_fill = style_red
   )
   three_stars_lightblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*\\*",
+    num_fmt = "#.#0 \\*\\*\\*",
     font_color = style_black,
-    bgFill = style_lightblue
+    bg_fill = style_lightblue
   )
   three_stars_midblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*\\*",
+    num_fmt = "#.#0 \\*\\*\\*",
     font_color = style_black,
-    bgFill = style_midblue
+    bg_fill = style_midblue
   )
   three_stars_darkblue <- openxlsx2::create_dxfs_style(
-    numFmt = "#.#0 \\*\\*\\*",
+    num_fmt = "#.#0 \\*\\*\\*",
     font_color = style_black,
-    bgFill = style_darkblue
+    bg_fill = style_darkblue
   )
 
   # create openxlsx2 workbook
@@ -237,50 +237,46 @@ cormatrix_to_excel <- function(data,
   # wb$styles_mgr$dxf
 
   # create the worksheets and write the data to the worksheets.
-  wb$add_worksheet("r_values")$add_data(x = cm)
-  wb$add_worksheet("p_values")$add_data(x = pf)
+  wb$add_worksheet("r_values")$add_data(x = cormatrix)
+  wb$add_worksheet("p_values")$add_data(x = p_val)
 
   # create conditional formatting for the stars (as well as colours as we have no)
   # one star
+  # Compute the cell range for styling
+  dims_fmt <- openxlsx2::wb_dims(cols = all.columns, rows = all.columns)
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= .2, r_values!B2 > 0, p_values!B2 < .05)",
     style = "one_star_pink"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= .2, p_values!B2 < .05)",
     style = "one_star_peach"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= .4, p_values!B2 < .05)",
     style = "one_star_red"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= -.2, r_values!B2 < 0, p_values!B2 < .05)",
     style = "one_star_lightblue"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= -.2, p_values!B2 < .05)",
     style = "one_star_midblue"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= -.4, p_values!B2 < .05)",
     style = "one_star_darkblue"
   )
@@ -288,43 +284,37 @@ cormatrix_to_excel <- function(data,
   # two stars
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= .2, r_values!B2 > 0, p_values!B2 < .01)",
     style = "two_stars_pink"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= .2, p_values!B2 < .01)",
     style = "two_stars_peach"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= .4, p_values!B2 < .01)",
     style = "two_stars_red"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= -.02, r_values!B2 < 0, p_values!B2 < .01)",
     style = "two_stars_lightblue"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= -.2, p_values!B2 < .01)",
     style = "two_stars_midblue"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= -.4, p_values!B2 < .01)",
     style = "two_stars_darkblue"
   )
@@ -332,43 +322,37 @@ cormatrix_to_excel <- function(data,
   # three stars
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= .2, r_values!B2 > 0, p_values!B2 < .001)",
     style = "three_stars_pink"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= .2, p_values!B2 < .001)",
     style = "three_stars_peach"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= .4, p_values!B2 < .001)",
     style = "three_stars_red"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 >= -.2, r_values!B2 < 0, p_values!B2 < .001)",
     style = "three_stars_lightblue"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= -.2, p_values!B2 < .001)",
     style = "three_stars_midblue"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 <= -.4, p_values!B2 < .001)",
     style = "three_stars_darkblue"
   )
@@ -376,61 +360,59 @@ cormatrix_to_excel <- function(data,
   # Other formatting
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(r_values!B2 = 1)",
     style = "gray_style"
   )
   wb$add_conditional_formatting(
     "r_values",
-    cols = all.columns,
-    rows = all.columns,
+    dims = dims_fmt,
     rule = "AND(p_values!B2 >= .05)",
     style = "no_star"
   )
 
   # p-values
-  wb$add_conditional_formatting("p_values",
-    cols = all.columns,
-    rows = all.columns,
+  wb$add_conditional_formatting(
+    "p_values",
+    dims = dims_fmt,
     rule = "< 10",
     style = "p_style"
   )
-  wb$add_conditional_formatting("p_values",
-    cols = all.columns,
-    rows = all.columns,
+  wb$add_conditional_formatting(
+    "p_values",
+    dims = dims_fmt,
     rule = "< .05",
     style = "p_style1"
   )
-  wb$add_conditional_formatting("p_values",
-    cols = all.columns,
-    rows = all.columns,
+  wb$add_conditional_formatting(
+    "p_values",
+    dims = dims_fmt,
     rule = "< .01",
     style = "p_style2"
   )
-  wb$add_conditional_formatting("p_values",
-    cols = all.columns,
-    rows = all.columns,
+  wb$add_conditional_formatting(
+    "p_values",
+    dims = dims_fmt,
     rule = "< .001",
     style = "p_style3"
   )
-  wb$add_conditional_formatting("p_values",
-    cols = all.columns,
-    rows = all.columns,
+  wb$add_conditional_formatting(
+    "p_values",
+    dims = dims_fmt,
     rule = "== 0",
     style = "gray_style"
   )
 
   ## Freeze Panes
-  wb$freeze_pane("r_values", firstCol = TRUE, firstRow = TRUE)
-  wb$freeze_pane("p_values", firstCol = TRUE, firstRow = TRUE)
+  wb$freeze_pane("r_values", first_col = TRUE, first_row = TRUE)
+  wb$freeze_pane("p_values", first_col = TRUE, first_row = TRUE)
 
   # Save Excel
   cat(paste0(
     "\n\n [Correlation matrix '", filename,
     ".xlsx' has been saved to working directory (or where specified).]"
   ))
-  openxlsx2::wb_save(wb, path = paste0(filename, ".xlsx"), overwrite = TRUE)
+  openxlsx2::wb_save(wb, file = paste0(filename, ".xlsx"), overwrite = TRUE)
 
   # open in Excel
   openxlsx2::xl_open(paste0(filename, ".xlsx"))