+++ tests/testthat/test-correlation.R
@@ -101,7 +101,6 @@ test_that("comparison with other packages", {
 })
 
 
-
 # Size
 test_that("format checks", {
   skip_if_not_or_load_if_installed("psych")
@@ -177,7 +176,7 @@ test_that("specific types", {
     y = as.ordered(sample(letters[1:5], 20, TRUE))
   )
 
-  correlation(data, method = "polychoric")
+  expect_warning(correlation(data, method = "polychoric"), regex = "It seems like")
 })
 
 test_that("correlation doesn't fail when BFs are NA", {
@@ -197,4 +196,44 @@ test_that("as.data.frame for correlation output", {
 
   set.seed(123)
   expect_snapshot(as.data.frame(correlation(ggplot2::msleep)))
+})
+
+test_that("correlation output with zap_small", {
+  set.seed(123)
+  d <- data.frame(
+    x = rnorm(10000),
+    y = rnorm(10000),
+    z = rnorm(10000)
+  )
+  r <- correlation::correlation(d)
+
+  expect_snapshot(summary(r), variant = "windows")
+  expect_snapshot(summary(r, zap_small = FALSE), variant = "windows")
+})
+
+
+test_that("missing values", {
+  data <- mtcars
+
+  data[1,1] <- NA
+
+  r_pairwise <- stats::cor(data[,1:5], use = "pairwise")
+  r_complete <- stats::cor(data[,1:5], use = "complete")
+
+  corr_pairwise <- correlation(data[,1:5])
+  corr_complete <- correlation(data[,1:5], missing = "keep_complete")
+
+  expect_equal(as.matrix(corr_pairwise), r_pairwise)
+  expect_equal(as.matrix(corr_complete), r_complete)
+
+
+
+  r_pairwise <- stats::cor(data[,1:2], data[,3:5], use = "pairwise")
+  r_complete <- stats::cor(data[,1:2], data[,3:5], use = "complete")
+
+  corr_pairwise <- correlation(data[,1:2], data[,3:5])
+  corr_complete <- correlation(data[,1:2], data[,3:5], missing = "keep_complete")
+
+  expect_equal(as.matrix(corr_pairwise), r_pairwise)
+  expect_equal(as.matrix(corr_complete), r_complete)
 })