+++ tests/testthat/test-display_print_dataframe.R
@@ -1,11 +1,8 @@
 test_that("display and print method works - markdown", {
   skip_on_cran()
-  skip_if(getRversion() < "4.0.0")
-  skip_if_not_or_load_if_installed("gt")
-
-  expect_snapshot(print(correlation(iris)))
-
+  skip_if_not_or_load_if_installed("knitr")
   expect_snapshot(display(correlation(iris)))
+  expect_snapshot(print_md(correlation(iris)))
 })
 
 # display and print method works - HTML -----------------------------
@@ -15,7 +12,6 @@ test_that("display and print method works - HTML", {
   skip_if(getRversion() < "4.0.0")
   skip_if_not_or_load_if_installed("gt")
 
-  expect_snapshot(display(print(correlation(subset(mtcars, select = c("wt", "mpg"))), format = "html")))
-
-  expect_snapshot(print(correlation(subset(mtcars, select = c("wt", "mpg"))), format = "html"))
+  expect_s3_class(display(correlation(subset(mtcars, select = c("wt", "mpg"))), format = "html"), "gt_tbl")
+  expect_s3_class(print_html(correlation(subset(mtcars, select = c("wt", "mpg")))), "gt_tbl")
 })