+++ tests/testthat/test-misc.R
@@ -1,7 +1,7 @@
 test_that("cor_to_cov", {
   cor <- cor(iris[1:4])
   cov <- cov(iris[1:4])
-  cov2 <- cor_to_cov(cor, var = sapply(iris[1:4], var))
+  cov2 <- cor_to_cov(cor, variance = sapply(iris[1:4], var))
   expect_equal(max(cov - cov2), 0, tolerance = 0.0001)
 })